Create encMenu.py
Used the dec menu so were on the same page. The process is in an infinite loop, there is a flag so the code knows when we are encrypting for the first time. This is so we only ask for plaintext once. The cipher name, key, and ciphertext are all put into a list, if the user encrypts again we use the last item in the list which is the most recent ciphertext. At the end the user is asked if they wish to encrypt again, if yes flag is marked true so the code knows this is not the first encryption and continues, if no a format for how the information will be displayed is printed and the list with the names, keys, and ciphertexts are output in the format I have the code print beforehand, then the code breaks from the infinite loop.